Add verbose configuration flags
This adds a config that can be enabled using the option
`--config=verbose`. It it is off by default, keeping the previous
behavior intact. When enabled, it makes build and test runs produce a
modest increase in diagnostic output. This is useful when debugging
GitHub Actions workflows.

Also in this commit: some comments acting as section dividers to make
navigating this file easier. (I find it helpful when trying to figure
out where a given change should go.)