Rename and improve dockertest.yml
This renames `dockertest.yml` to `ci_docker_tests.yaml`, to make the
purpose more clear from the file name. In addition, it adds a step to
check whether the changes in a PR are such that trying to build Docker
images is warranted. (Specifically, if the only changes involve
documentation files, then there's no point in running the Docker tests.)