Start a pyproject.toml file
The initial contents are settings for cibuildwheel and pytest. The
cibuildwheel settings came from the GitHub Actions workflow for
building wheels. Putting the settings here means the settings are
available in more situations, such as when running cibuildwheel
manually for testing, or if we create more scripts or workflows that
run cibuildwheel.