Check cpu features more portably in test_libs.sh
The previous commands to get cpu features on Windows failed on GitHub
Windows-2022 runners. It seems the underlying commands used by the
script changed. The new approach in this commit uses the Python
`py-cpuinfo` package, which works on Windows and has the bonus of
working on Mac and Linux too, allowing us to simply the test_libs.sh
further.