Make "bazel build foo" without options default to basic build
The comments in `.bazelrc`, and the instructions in `docs/bazel.md`,
implied that the default build and test behavior would be the "basic"
build, without AVX, SSE, or OpenMP options. However, this was not the
case in practice: if no options were given, it would still end up
build with the AVX and SSE options. This happened because the BUILD
file isn't set up to add things only when options are used; it's set
up to build everything by default.

The recent addition of Bazel tags to the BUILD file now allows us to
remedy this. The changes in this commit change the default `build` and
`test` actions be exclude things tagged with "avx" and "sse". If the
user adds `--config=avx`, then the "avx" tag are added back;
similarly, specifying `--config=sse` adds the "sse" tag.

This now makes the behavior consistent with the documentation, and
also has the advantage that someone blindly doing a "bazel test
tests:all" without any options will not encounter failures on systems
that don't have AVX features.