Adjust and spiff up README format
The way sections were arranged in the REAMDE file left lacked
hierarchical consistency. In addition, there was not much of an
introduction, which could leave new visitors a bit confused. I
made the section structure more hierarchical, moved a section or two
around for better logical flow, and added the usual badges and other
visual elements at the top.