Update CMake files to make CUDA builds work
The bulk of these fixes were originally done by @fdmalone. This
updates the CMake rules to avoid deprecated constructs for finding
the CUDA rules and configurations. In addition, while we're at it,
this also updates the minimum Python version to 3.10 so that this is
consistent with the requirements defined elsewhere in the latest
version of qsim.