Don't add CMAKE_CUDA_COMPILER flag if don't find nvcc
Unconditionally adding CMAKE_CUDA_COMPILER to cmake_args confuses the
CMake CUDA support detection code when a build is running inside a
Docker container. The problems seems to be that when the pybind11
building stage is reached, a nonzero value for CMAKE_CUDA_COMPILER
makes CMake's CMakeCUDAFindToolkit.cmake module think that CUDA is
available, but then it can't find nvcc inside the Docker container.

The error is evident when running `docker-compose build` inside the
subdirectory `install/tests`. The error is:

```
-- pybind11 v2.13.6
  -- Found PythonInterp: /qsim/test_env/bin/python3 (found suitable
     version "3.12.3", minimum required is "3.7")
  -- Found PythonLibs: /usr/lib/x86_64-linux-gnu/libpython3.12.so
  -- Performing Test HAS_FLTO
  -- Performing Test HAS_FLTO - Success
  -- qsim: c
  -- Found Python3: /qsim/test_env/bin/python3.12 (found suitable
     version "3.12.3", minimum required is "3.10") found components:
     Interpreter Development Development.Module Development.Embed
  CMake Error at
  /usr/local/share/cmake-3.31/Modules/Internal/CMakeCUDAFindToolkit.cmake:104
  (message):
    Failed to find nvcc.

    Compiler requires the CUDA toolkit.  Please set the
    CUDAToolkit_ROOT
    variable.
  Call Stack (most recent call first):
    /usr/local/share/cmake-3.31/Modules/CMakeDetermineCUDACompiler.cmake:85
    (cmake_cuda_find_toolkit)
    pybind_interface/cuda/CMakeLists.txt:16 (project)

```