Update CMake files for CUDA compatibility on Linux
Testing things on a Linux system with an actual GPU and NVIDIA's CUDA
library revealed some problems in finding the CUDA compiler and other
components. The most important is to use

```
include(CheckLanguage)
check_language(CUDA)
```

to check for CUDA support now.

In adition, while at it, I updated the files further for MacOS (e.g.,
to make use of `CMAKE_APPLE_SILICON_PROCESSOR`) and to add missing
copyright statements.