Correct the default Bazel vector instructions options
The documented way that `bazel build`, `bazel test`, and similar
commands behave without options like `--config=avx` is that the AVX,
SSE, and other optimization options are not applied. However, this was
not what was happening: the default ended up applying them. The
previous changes to `.bazelrc` and `dev_tools/test_libs.sh` that I did
in PR #758 and other recent PRs not fully address this.

The changes here invert the way `--build_tag_filters` and
`--test_tag_filters` are used: they are set to filter out things
tagged with `avx` and `sse` by default in `.bazelrc`, and options like
`--config=sse` add the things back.

To make this work, `dev_tools/test_libs.sh` has to be updated as well.