Suppress CMake version warning coming from Pybind11
When building using cibuildwheel, the following warning is printed:

```
 CMake Deprecation Warning at .../pybind11-src/CMakeLists.txt:13
 (cmake_minimum_required):
    Compatibility with CMake < 3.10 will be removed from a future
    version of
    CMake.

    Update the VERSION argument <min> value.  Or, use the
    <min>...<max> syntax to tell CMake that the project requires
    at least <min> but has been updated to work with policies
    introduced by <max> or earlier.

```

This is coming from Pybind11's CMake files. We can't do anything about
it, and our devs end up wasting time looking for the cause in qsim's
files.

The change here suppresses deprecation warnings around the part where
Pybind11 is included.

This change should be reverted when Pybind11 updates the minimum
required CMake version to a number that doesn't trigger the warning.