Add Bazel config options for SSE & AVX
Although the recent changes made sure not to build & test the versions
of files for AVX & SSE on systems that lacked those features, the
converse was not done: it did not add `--config=avx` and/or
`--config=sse` if the current host supported them.

This commit adds the necessary changes to do that.