Fix comparison to workflow_dispatch inputs
The behavior of GitHub Actions workflows is counterintuitive (at least
to me). When the trigger event is not `workflow_dispatch` and thus the
input variable `bazel-cache` is not set, the expression
`inputs.bazel-cache != false` turns out to be `false`. The reason is
that `inputs.bazel-cache` is `null`, and in the JavaScript-like
comparisons used by GitHub Actions, `null` is not strictly equal to
`false`.

The readable solution seems to be to test the full path to input
variables.

For better code readability, I've set the value to a environment
variable.