Install dev-requirements.txt into Python environments
After switching to Python 3.12 for the Python environment in the
workflows, one of the workflows started failing. That workflow only
loaded `requirements.txt`. It wasn't clear why, and took time to
debug, and ultimately I think the problem was elsewhere. However,
what wasted my time is chasing a red herring thinking that the problem
was due to not loading `dev-requirements.txt`.

I think it's better if all the workflows load both `requirements.txt`
and `dev-requirements.txt` to reduce the number of variables in future
debugging episodes. (They may not all need it, and if there's a
specific reason for a workflow not to load dev-requirements then
that's okay. It's just that it's better if the default is uniform
across all the workflows.)