Add a .dockerignore file
The configuration in `install/tests/Dockerfile` copies the entire qsim
source directory into the Docker container. If the user previously did
a `pip install .` in the qsim source directory, this will copy the egg
directory too. This confuses pip installation commands inside the
Docker environment.

Adding a `.dockerignore` file to ignore the egg directory solves this,
and is more maintainable than trying to copy only select parts of the
qsim source directory.