Handle comma-separated CUDAARCHS value
In at least one of our internal team instructions, an example was
given with `CUDAARCHS` using comma-separated values instead fo
semicolon-separated. That's not the correct syntax and if left that
way, the error that results is difficult to diagnose. Let's save
ourselves the future hassle by checking for it and telling the user.