Detect AVX & SSE and only build corresponding parts
Previously, the CMake configuration was such that it would always
build the AVX2, AVX512 & SSE2 Pybind11 modules without testing whether
the current system supported those options. The changes here use CMake
features to detect whether the architectural features are in fact
available, and only attempt to build the appropriate modules.

In addition, there is a minor bit of refactoring to group some of the
code more logically, and to add some more informational message
printouts.