Work around Pillow build failures on Linux
The work around works like this: we pip-install a version of Pillow <
11.3 using the `before-test` option of cibuildwheel, so that when
cibuildwheel does its pip-install of the qsimcirq package that it
built, the Pillow dependency is satisfied. This works because the
requirement of Pillow in Matplotlib does not require 11.3; it just
requires a version higher than 9, which our pre-installation of Pillow
satisfies.

This should be a temporary workaround only. We should update qsim to
use cibuildwheel 3 and manylinux_2_28 – this problem may go away at
that point.