Rewrite the logic and remove unnecessary CMake code
Due to the fact that the top-level `CMakeLists.txt` file includes
these files using `add_directory`, it's not necessary for them to do
things like include `GetPybind.cmake` or call `project()`. Those
directives are already done by the top-level file. We can simplify
these sub-`CMakeLists.txt` files.

In addition, this commit rewrites the logic for determining which flag
to pass to the compiler on Windows. The Windows case is complicated
due to changes in the flags in MSVC over the years. The new logic
tries to be robust in the face of different possibilities.