Remove unnecessary CMake code and simply setting the arch
Due to the fact that the top-level `CMakeLists.txt` file includes
these files using `add_directory`, it's not necessary for them to do
things like include `GetPybind.cmake` or call `project()`. Those
directives are already done by the top-level file. We can simplify
these sub-`CMakeLists.txt` files.

In addition, the logic for setting `CMAKE_CUDA_ARCHITECTURES` can be
made very simple since we don't do complicated things to try to figure
out which CUDA/GPU variants the host has. If the `CUDAARCHS`
environment variable is not set and `CMAKE_CUDA_ARCHITECTURES` is not
set either, we just default to using `native` and hope for the best.