Revise workflows to fix GHA syntax & rename files
The values from workflow calls turn out _always_ to be strings,
even when you try to make sure to produce Booleans. This means
tests of values output by workflows called by other workflows
have to be of the form `${{variable == 'true'}}` rather than
(e.g.) `${{variable}}` or even `${{variable == true}}`. This is
poorly documented in GitHub's documentation, IMHO.

Along with fixing the conditionals, this PR renames the reusable
workflow files to start with leading underscore characters. That
seems to be more conventional in other GitHub repos.