Rework workflows for better status reporting
This uses a different strategy: uses a separate job in each workflow to
report the results. This makes it possible to handle skipping smoothly
yet still allow jobs to work as desired in CI checks. It also makes it
easier to select the jobs in the branch rules for running CI checks,
because there's only 1 of each kind to select (instead of every matrix
combination).