Remove screenshot from .github/SECURITY.md (#909)
It turns out there are different paths to get to the security page on
GitHub, and one of them ends up with a broken image. In retrospect, my
attempt to be newbie-friendly went overboard with this, and it was a bad
idea to try to provide a screenshot because it's too prone to breakage.
Let's remove it.