Add `.pylintrc`, update `dev-requirements.txt`, and address some trivial Pylint warnings (#911)
This adds a `.pylintrc` file based on the one from Cirq, with a change
to the line length (80 for this project) and removal of one of the
checks because of a difference in conventions. It also adds `pylint` to
`dev-requirements.txt` (using the same version that Cirq uses). Finally,
it also adjusts a few Python files to address consequent linter
warnings.