Fix #884: test for `nvidia-smi`, not `nvcc`
It seems that checking for the program `nvidia-smi` is the more reliable
way to determine if a system has a usable NVIDIA GPU and the necessary
drivers installed for running CUDA applications. Checking for `nvcc` may
only tell you if the system can compile CUDA code, not necessarily if it
can run it.