Fix #935: warning in gate_appl.h
The problem seems to be that `discarded_results` in `ApplyFusedGate` is
passed by value to the inner `ApplyFusedGate`. When
`std::move(measure_result)` happens inside, it's moving from a stack
object into a vector that is also on the stack (because it's a parameter
passed by value). Then when the inner `ApplyFusedGate` returns, its copy
of `mresults` is destroyed, triggering the warning.

The proper fix may be to change the calling function (the one on lines
221-228) to not use a temporary vector that goes out of scope. Since the
results are discarded anyway, they don't need to be stored in a vector
at all. We can modify the main `ApplyFusedGate` (lines 191-211) to not
add the result to the vector if it's not needed. However, that requires
a more significant change. A simpler fix for the warning is to just
heap-allocate the `discarded_results` vector in the overload function.