Fix #935: avoid the need to dealocate an object
The problem that lead to the compiler warnings is that
`discarded_results` in `ApplyFusedGate` is passed by value to the inner
`ApplyFusedGate`. When `std::move(measure_result)` happens inside, it's
moving from a stack object into a vector that is also on the stack
(because it's a parameter passed by value). Then when the inner
`ApplyFusedGate` returns, its copy of `mresults` is destroyed,
triggering the warning.

The solution used here is to refactor the logic into a separate
implementation function that accepts a pointer to the results vector.
This pointer can be `nullptr` to explicitly signal that measurement
results should be discarded. This change eliminates the temporary
`discarded_results` vector. Since the vector is never created, its
destructor is never called, and the compiler warning vanishes. This is
also faster, as you save the cost of constructing and destructing that
temporary vector.