Fix #930: add a `.gemini/styleguide.md` for Gemini Code Assist (#931)
Add a guide for Gemini Code Assist to teach it about the conventions
used in this project.

[Gemini Code Assist](https://codeassist.google/) is a coding assistant
based on Google Gemini. Among other things, it has an [integration with
GitHub](https://developers.google.com/gemini-code-assist/docs/set-up-code-assist-github).
Google has enabled it in our organization. It acts as a [code reviewer
for pull
requests](https://developers.google.com/gemini-code-assist/docs/review-github-code).
Its behavior can be
[customized](https://developers.google.com/gemini-code-assist/docs/customize-gemini-behavior-github)
to some extent, and it can be informed about the practices and
conventions used in a project through the use of a [styleguide.md
file](https://developers.google.com/gemini-code-assist/docs/customize-gemini-behavior-github#styleguide.md).

The `styleguide.md` file in this PR is essentially a draft contributors'
guide for the qsim project. The contents can serve as a starting point
for adding a section to the qsim documentation.

---------

Co-authored-by: Pavol Juhas <pavol.juhas@gmail.com>