Don't run Scorecard and OSV on merge queue events (#957)
When Scorecard runs on merge events in the merge queue, it seems to
produce an error: "Only the default branch main is supported". I don't
fully understand why, because the `merge_queue` explicitly had a branch
condition to only run on `main`. (Maybe the error is not about `main`,
but about the fact that it's `main` in someone's fork?)

However, this made me realize that it's not essential to run Scorecard
and OSV on merge queue events because they've already been run on the
pull request, and in addition, we do regular scheduled execution over
the whole code base.