Detectd missing cpuinfo package and let the user know
Due to the fact that the package name is not the same as the module name
(`py-cpuinfo` vs `cpuinfo`), the error that results from trying to run
this script when `py-cpuinfo` is not installed can be confusing. Let's
help users & developers by testing it and pointing them in the right
direction.