Adjust GitHub's file statistics
IMHO it's useful to see Markdown files as one of the types of files
listed on the GitHub front page in the file statistics section,
alongside other files. It gives a sense for the amount of documentation
present in a project.

On the other hand, some configuration files are not useful to count,
such as .md files in a `.gemini/` subdirectory and configuration files
for tools like git.